Update to Python 3
Try to understand encoding was too hard... In this fix, I tried to
update Python 2 code to Python 3. But Windows is having some issues...
so I'm updating the script as best I can. Then, I'm going to go back in
and update it from a dual-booted Linux system. Stay tuned for
stability....